Update CM104.sql
Hi @sabetiv - Index Data is still looking into whether inventory.creation_date is correct or not, but I would like your advice on the parentheses. Currently, they seem to only be querying and returning results from the 994. We need to be querying both the 994 and 998, but I seem to be at the point where I'm running my head into a wall. Thanks for any help!